backend/lna-aggregators/pyproject.toml
[tool.poetry]
name = "lna-aggregators"
version = "0.1.0"
description = "Aggregators module for LNA"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "lna_aggregators" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
google-genai = "^1.13.0"
google-api-core = "^2.24.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
lna-db = { path = "../lna-db", develop = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
